Move all the reactive functions to RxSession
