+++ hibernate-rx-core/src/main/java/org/hibernate/rx/RxHibernateSession.java
@@ -24,6 +24,4 @@ public interface RxHibernateSession extends Session {
 
 	RxActionQueue getRxActionQueue();
 
-    CompletionStage<Void> rxFlush();
-
 }
