+++ hibernate-rx-core/src/main/java/org/hibernate/rx/impl/RxHibernateSessionImpl.java
@@ -63,98 +63,6 @@ public class RxHibernateSessionImpl extends SessionDelegatorBaseImpl implements 
 		return this;
 	}
 
-//	private void firePersist(PersistEvent event) {
-//		try {
-////			checkTransactionSynchStatus();
-////			checkNoUnresolvedActionsBeforeOperation();
-//
-//			for ( PersistEventListener listener : listeners( EventType.PERSIST ) ) {
-//				listener.rxOnPersist( event );
-//			}
-//		}
-//		catch (MappingException e) {
-//			throw exceptionConverter.convert( new IllegalArgumentException( e.getMessage() ) );
-//		}
-//		catch (RuntimeException e) {
-//			throw exceptionConverter.convert( e );
-//		}
-//		finally {
-////			try {
-////				checkNoUnresolvedActionsAfterOperation();
-////			}
-////			catch (RuntimeException e) {
-////				throw exceptionConverter.convert( e );
-////			}
-//		}
-//	}
-
-	@Override
-	public void flush() {
-		checkOpen();
-		doFlush();
-	}
-
-	@Override
-	public CompletionStage<Void> rxFlush() {
-	    checkOpen();
-	    return doFlush();
-	}
-
-	private CompletionStage<Void> doFlush() {
-//		checkTransactionNeeded();
-//		checkTransactionSynchStatus();
-
-//			if ( persistenceContext.getCascadeLevel() > 0 ) {
-//				throw new HibernateException( "Flush during cascade is dangerous" );
-//			}
-
-		    CompletionStage<Void> ret = RxUtil.nullFuture();
-			FlushEvent flushEvent = new FlushEvent( this );
-			for ( FlushEventListener listener : listeners( EventType.FLUSH ) ) {
-				ret = ret.thenCompose(v -> ((RxFlushEventListener)listener).rxOnFlush( flushEvent ));
-			}
-
-//			delayedAfterCompletion();
-			return ret.exceptionally(x -> {
-			    if(x instanceof RuntimeException)
-			        throw exceptionConverter.convert( (RuntimeException)x );
-			    else
-			        return RxUtil.rethrow(x);
-			});
-	}
-
-	private <T> Iterable<T> listeners(EventType<T> type) {
-		return eventListenerGroup( type ).listeners();
-	}
-
-	private <T> EventListenerGroup<T> eventListenerGroup(EventType<T> type) {
-		return getFactory().getServiceRegistry().getService( EventListenerRegistry.class ).getEventListenerGroup( type );
-	}
-
-	private void managedFlush() {
-		if ( isClosed() ) {// && !waitingForAutoClose ) {
-//			log.trace( "Skipping auto-flush due to session closed" );
-			return;
-		}
-//		log.trace( "Automatically flushing session" );
-		doFlush();
-	}
-
-	@Override
-	public void flushBeforeTransactionCompletion() {
-		final boolean doFlush = isTransactionFlushable()
-				&& getHibernateFlushMode() != FlushMode.MANUAL;
-
-		try {
-			if ( doFlush ) {
-				managedFlush();
-			}
-		}
-		catch (RuntimeException re) {
-			throw exceptionMapper.mapManagedFlushFailure( "error during managed flush", re, this );
-		}
-	}
-
 	private boolean isTransactionFlushable() {
 		if ( getCurrentTransaction() == null ) {
 			// assume it is flushable - CMT, auto-commit, etc
